OPENEUROPA-1503: Move custom behat steps to separate context.
