+++ behat.yml.dist
@@ -6,6 +6,7 @@ default:
       contexts:
         - Drupal\DrupalExtension\Context\MinkContext
         - Drupal\DrupalExtension\Context\DrupalContext
+        - Drupal\Tests\oe_editorial\Behat\ContentLockContext
         - Drupal\Tests\oe_editorial\Behat\FeatureContext
         - OpenEuropa\Behat\TransformationContext:
             pages:
